Move Vet360 field/sections into vet360 directory and update references
