+++ src/applications/personalization/profile360/components/ProfileView.jsx
@@ -5,7 +5,7 @@ import DowntimeNotification, { externalServices, externalServiceStatus } from '.
 import DowntimeApproaching from '../../../../platform/monitoring/DowntimeNotification/components/DowntimeApproaching';
 import recordEvent from '../../../../platform/monitoring/record-event';
 
-import Vet360TransactionReporter from '../containers/Vet360TransactionReporter';
+import Vet360TransactionReporter from '../vet360/containers/Vet360TransactionReporter';
 
 import Hero from './Hero';
 import ContactInformation from './ContactInformation';
