+++ src/applications/personalization/profile360/vet360/components/AddressField/AddressForm.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
 
-import { focusElement } from '../../../../platform/utilities/ui';
+import { focusElement } from '../../../../../../platform/utilities/ui';
 
-import ErrorableSelect from '../../../letters/components/ErrorableSelect';
-import ErrorableTextInput from '../../../letters/components/ErrorableTextInput';
-import { STATE_CODE_TO_NAME, MILITARY_CITIES, MILITARY_STATES } from '../../../letters/utils/constants';
+import ErrorableSelect from '../../../../../letters/components/ErrorableSelect';
+import ErrorableTextInput from '../../../../../letters/components/ErrorableTextInput';
+import { STATE_CODE_TO_NAME, MILITARY_CITIES, MILITARY_STATES } from '../../../../../letters/utils/constants';
 
 /**
  * Input component for an address.
