+++ src/applications/personalization/profile360/vet360/components/base/Vet360EditModal.jsx
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 
 import Modal from '@department-of-veterans-affairs/formation/Modal';
 
-import Vet360EditModalErrorMessage from '../components/Vet360EditModalErrorMessage';
-import LoadingButton from '../components/LoadingButton';
-import Vet360EditModalActionButtons from '../components/Vet360EditModalActionButtons';
+import Vet360EditModalErrorMessage from './Vet360EditModalErrorMessage';
+import LoadingButton from './LoadingButton';
+import Vet360EditModalActionButtons from './Vet360EditModalActionButtons';
 
 export default class Vet360EditModal extends React.Component {
