+++ src/applications/personalization/profile360/vet360/containers/Vet360PendingTransactionCategory.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { connect } from 'react-redux';
 import AlertBox from '@department-of-veterans-affairs/formation/AlertBox';
 
-import { refreshTransaction } from '../actions/updaters';
+import { refreshTransaction } from '../../actions/updaters';
 
-import Vet360TransactionPending from '../components/Vet360TransactionPending';
+import Vet360TransactionPending from '../components/base/Vet360TransactionPending';
 
-import { TRANSACTION_CATEGORY_TYPES } from '../constants/vet360';
-import { selectVet360PendingCategoryTransactions } from '../selectors';
+import { TRANSACTION_CATEGORY_TYPES } from '../../constants/vet360';
+import { selectVet360PendingCategoryTransactions } from '../../selectors';
 
 function Vet360PendingTransactionCategory({ refreshTransaction: dispatchRefreshTransaction, transactions, hasPendingCategoryTransaction, categoryType, children }) {
   if (!hasPendingCategoryTransaction) return <div>{children}</div>;
